Multiple Statements - Check for duplicates
