+++ routes/api_v1.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\Api\v1\OpenSearchAPIController;
 use App\Http\Controllers\Api\v1\PlatformAPIController;
 use App\Http\Controllers\Api\v1\PlatformUserAPIController;
 use App\Http\Controllers\Api\v1\StatementAPIController;
+use App\Http\Controllers\Api\v1\StatementMultipleAPIController;
 use App\Http\Controllers\Api\v1\UserAPIController;
 use App\Http\Controllers\PlatformController;
 use Illuminate\Support\Facades\Route;
@@ -23,7 +24,7 @@ Route::middleware('auth:sanctum')->group(static function () {
     Route::get('statement/{statement}', [StatementAPIController::class, 'show'])->name('api.v1.statement.show')->can('view statements');
     Route::get('statement/existing-puid/{puid}', [StatementAPIController::class, 'existingPuid'])->name('api.v1.statement.existing-puid')->can('view statements');
     Route::post('statement', [StatementAPIController::class, 'store'])->name('api.v1.statement.store')->can('create statements');
-    Route::post('statements', [StatementAPIController::class, 'storeMultiple'])->name('api.v1.statements.store')->can('create statements');
+    Route::post('statements', [StatementMultipleAPIController::class, 'store'])->name('api.v1.statements.store')->can('create statements');
     Route::group(['middleware' => ['can:administrate']], static function () {
         Route::post('opensearch/search', [OpenSearchAPIController::class, 'search'])->name('api.v1.opensearch.search');
         Route::post('opensearch/count', [OpenSearchAPIController::class, 'count'])->name('api.v1.opensearch.count');
